urukul: remove injection support
After some more thinking I decided it is not worth it to include urukul injection support here, and should be done as another PR.

The previous "hidden experiment spawner" based approach worked very well, yet is flawed, not only that it will stall the UI thread, but it also requires direct access to the dashboard and bypassed the moninj proxy entirely, so it is better for users to submit an injection themselves via experiment.

However, there remaining frequency changer experiment can be reused, and we could include it for the user as demo, let them extend it themselves according to their needs.

Signed-off-by: Steve Fan <sf@m-labs.hk>